requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
psycopg-binary==3.0.15
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
unicorn==2.0.0
uvicorn==0.18.2
